+++ Backend/productService/src/main/java/com/micro/productservice/repository/FileRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface FileRepository extends JpaRepository<File, Long> {
+public interface FileRepository extends JpaRepository<File, Integer> {
 }
